.history/backend/requirements_20250214143349.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
python-dotenv==0.19.0
openai
matplotlib==3.7.1
numpy==1.24.3
torch==2.0.1
pillow==8.3.2
transformers==4.28.1
redis==4.6.0
python-redis==0.1.2
pino==0.6.0
python-json-logger==2.0.7
scipy==1.7.3
